% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvarpyp.R
\name{cvarpyp}
\alias{cvarpyp}
\title{Clustering Variying Coefficients with a Pitman-Yor Process}
\usage{
cvarpyp(
  ID,
  W,
  X = NULL,
  Z,
  t,
  Y,
  num_knots = 10,
  K_max,
  num_iters = 10000,
  thinning_unit = 5,
  nu = 1,
  lambda = 0.5,
  Psi = NULL,
  kappa = 10,
  sd_nu = 1,
  sd_lambda = 1,
  pi.a = 1,
  pi.b = 1,
  stick_act = c(1, 10),
  stick_nact = c(1, 10)
)
}
\arguments{
\item{ID}{Subjects' ID vector}

\item{W}{A part of design matrix for varying coefficients}

\item{X}{A part of design matrix for overall fixed effects}

\item{Z}{A part of design matrix for random effects}

\item{t}{A time vector}

\item{Y}{A binary response vector}

\item{num_knots}{The number of knots for splines}

\item{K_max}{The upper bound of the number of clusters for the stick-breaking process}

\item{num_iters}{The number of iterations for the partially collapsed Gibbs sampler}

\item{sampling_unit}{Thinning unit}
}
\description{
Clustering Variying Coefficients with a Pitman-Yor Process
}
