% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvarpyp.R
\name{fcvarpyp}
\alias{fcvarpyp}
\title{Clustering Variying Coefficients with a Pitman-Yor Process}
\usage{
fcvarpyp(
  ID,
  W,
  X = NULL,
  Z,
  t,
  Y,
  num_knots = 10,
  K_max,
  num_iters = 10000,
  thinning_unit = 5,
  nu = 1,
  lambda = 0.5,
  Psi = NULL,
  beta = NULL,
  kappa = 10,
  stick_act = c(1, 10),
  stick_nact = c(1, 10)
)
}
\arguments{
\item{ID}{A vector of subjects' IDs.}

\item{W}{A part of design matrix for varying coefficients}

\item{X}{A part of design matrix for time-invariant fixed effects}

\item{Z}{A part of design matrix for random effects}

\item{t}{A vector of time observations}

\item{Y}{A vector of binary responses}

\item{num_knots}{The number of knots for the basis splines}

\item{K_max}{The upper bound of the number of clusters for the stick-breaking process}

\item{num_iters}{The number of iterations for the partially collapsed Gibbs sampler}

\item{thinning_unit}{Thinning unit of the pcg sampler}

\item{nu}{An initial value of the concentration parameter}

\item{lambda}{An initial value of the discount parameter}

\item{Psi}{An initial covariance matrix of the random effects}

\item{beta}{An initial vector of the fixed effects}
}
\description{
Clustering Variying Coefficients with a Pitman-Yor Process
}
\examples{
library(cvarpyp)

#SimulationData is an example dataset.

ID <- SimulationData$ID
W <- SimulationData[c('W1','W2','W3','W4')]
X <- SimulationData[c('X1','X2','X3')]
Z <- SimulationData[c('Z1','Z2')]
t <- SimulationData$t
Y <- SimulationData$Y

output <- fcvarpyp(ID = ID,
               W = W,
               X = X,
               Z = Z,
               t = t,
               Y = Y,
               num_knots = 15,
               K_max = 10,
               num_iters = 10000)

plot(output$fixed_effect, position = 1)
plot(output$random_effect, row_position = 1, col_position = 1)
plot(output$cluster)
plot(output$varying_coefficient, cluster_number = 5, variable_number = 1)
plot(output$latent_location, cluster_number = 5, variable_number = 1, time_range = output$time_range, knot_position = output$knot_position)
plot(output$nu)
plot(output$lambda)

}
