\name{fvcc-package}
\alias{fvcc-package}
\alias{fvcc}
\docType{package}
\title{
\packageTitle{fvcc}
}
\description{
\packageDescription{fvcc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{fvcc}
\packageIndices{fvcc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{fvcc}

Maintainer: \packageMaintainer{fvcc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
