% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_simulation_data}
\alias{get_simulation_data}
\title{Get simulation data}
\usage{
get_simulation_data(sbj1, sbj2, sbj3, seed = 0)
}
\arguments{
\item{sbj1}{The number of subjects of the first group}

\item{sbj2}{The number of subjects of the second group}

\item{sbj3}{The number of subjects of the third group}
}
\description{
Get simulation data
}
