% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{plot.cluster}
\alias{plot.cluster}
\title{Draw the posterior results of clusters}
\usage{
\method{plot}{cluster}(cluster.sample, burn = NULL)
}
\description{
Draw the posterior results of clusters
}
