% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{plot.fixed_effect}
\alias{plot.fixed_effect}
\title{Draw the posterior results of fixed effects}
\usage{
\method{plot}{fixed_effect}(fixed_effect_samples, position, burn = NULL)
}
\description{
Draw the posterior results of fixed effects
}
