% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{plot.lambda}
\alias{plot.lambda}
\title{Draw the posterior histogram of the discount parameter}
\usage{
\method{plot}{lambda}(lambda_sample, burn = NULL)
}
\description{
Draw the posterior histogram of the discount parameter
}
