% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{plot.latent_location}
\alias{plot.latent_location}
\title{Draw the posterior results of latent location parameters}
\usage{
\method{plot}{latent_location}(
  latent_location_samples,
  cluster_number,
  variable_number,
  time_range,
  knot_position,
  burn = NULL
)
}
\description{
Draw the posterior results of latent location parameters
}
