% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{plot.nu}
\alias{plot.nu}
\title{Draw the posterior histogram of the concentration parameter}
\usage{
\method{plot}{nu}(nu_sample, burn = NULL)
}
\description{
Draw the posterior histogram of the concentration parameter
}
