% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{plot.random_effect}
\alias{plot.random_effect}
\title{Draw the posterior results of random effects}
\usage{
\method{plot}{random_effect}(random_effect_samples, row_position, col_position, burn = NULL)
}
\description{
Draw the posterior results of random effects
}
