% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-visualization.R
\name{plot.varying_coefficient}
\alias{plot.varying_coefficient}
\title{Draw the posterior results of varying coefficients}
\usage{
\method{plot}{varying_coefficient}(varying_coefficient_samples, cluster_number, variable_number)
}
\description{
Draw the posterior results of varying coefficients
}
