% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{get_simulation_data}
\alias{get_simulation_data}
\title{Get simulation data}
\usage{
get_simulation_data(N, K, M, seed = 0)
}
\arguments{
\item{N}{The number of subjects}

\item{K}{The number of clusters}

\item{M}{The average number of observations}
}
\description{
Get simulation data
}
